Version 0.1.3: Reduced the keepalive to 10 minutes, bug fixes

+++ index.js
@@ -154,7 +154,7 @@ function mqttConnect() {
             username: simId,
             password: PRIVATE.mqttToken,
             clientId: simId,
-            keepalive: 60*60,
+            keepalive: 60*10,
             clean: false,
             reconnectPeriod: 1000 * 60 * 10
         }
@@ -216,7 +216,7 @@ function openTunnel(queenPort, antPort, target) {
         });
         // if no error after SSH_TIMEOUT 
         setTimeout(function(){reject({process: myProcess, msg:"SSH timeout"}); }, SSH_TIMEOUT);
-    })
+    });
 
 }
 
@@ -373,7 +373,9 @@ function commandHandler(fullCommand, sendFunction, topic) { // If a status is se
                     break;
                 case 'reboot':               // Reboot the system
                     sendFunction(topic, JSON.stringify({command: command, result: 'OK'}));
-                    spawn('reboot');
+                    setTimeout(function () {
+                        spawn('reboot');
+                    }, 1000);
                     break;
                 case 'resumerecord':         // Start recording
                     wifi.record(MEASURE_PERIOD);
@@ -386,7 +388,8 @@ function commandHandler(fullCommand, sendFunction, topic) { // If a status is se
                     sendFunction(topic, JSON.stringify({command: command, result: 'OK'}));
                     break;
                 case 'closetunnel':          // Close the SSH tunnel
-                    sshProcess.kill('SIGINT');
+                    if (sshProcess)
+                        sshProcess.kill('SIGINT');
                     setTimeout(function () {
                         if (sshProcess)
                             sshProcess.kill();
